requirements.txt
flask
requests
python-dotenv
requests 
urllib3